This should add good comments and y-axis
We will continue to do in html javascript until the project is finished. Then we will try again to transition to a separate javascript file. This will need to be very well commented in order for this to be successful. The intended goal will be to use the canvas width and height to our advantage to perform a ton of hand calculations for this interface as well as functions to help us plot the simple graph values from 0 to 2pi in pi/2 intervals. We will see if we can get multiple periods of this graph for visual purposes.