Add modularity to the vertical gridline creation
The goal is by the end of this extremely long and complex feature is to have a robust graph that can fit from any given x limits into the canvas, so that there is no need to do any scrolling at all. This feature will be implemented in many different stages and will require very careful and detailed documentation in order to account for all possible user inputs for the x limits.